rename 'web_poet.overrides' into 'web_poet.rules'
